refactor denudation code
