+++ .entangled/filedb.json
@@ -1,5 +1,5 @@
 {
-  "version": "2.0.2",
+  "version": "2.0.5",
   "files": [
     {
       "path": "docs/src/carbocat-transport.md",
@@ -14,13 +14,6 @@
       "modified": "2024-06-10T14:56:12.712671",
       "hexdigest": "c6615334f9a5d74bd3b7f3410ed6ee783d1c244e1f3b35c5a4e0e66ab1570b74",
       "size": 1202
-    },
-    {
-      "path": "docs/src/bs92-w-erosion.md",
-      "deps": null,
-      "modified": "2024-05-27T11:48:20.660583",
-      "hexdigest": "5dcd54a597f93820adb57acfa3cb365be4346226182f9cfaf9f7bc462faab789",
-      "size": 1285
     },
     {
       "path": "docs/src/bosscher-1992.md",
@@ -108,15 +101,6 @@
       "modified": "2024-04-25T15:48:30.545337",
       "hexdigest": "7f642540e05f33ecb649d42dd6274f1469096c51b43e318066ddd6f5e8f6cd0d",
       "size": 522
-    },
-    {
-      "path": "src/Erosion.jl",
-      "deps": [
-        "docs/src/bs92-w-erosion.md"
-      ],
-      "modified": "2024-05-27T11:48:20.698583",
-      "hexdigest": "7a49c1fb19be436425941b9ec94ee2e5b316faa7a7b71bd301d77ffdfc3dcde8",
-      "size": 1223
     },
     {
       "path": "src/CaProd.jl",
@@ -522,49 +506,48 @@
   ],
   "source": [],
   "target": [
-    "src/Visualization.jl",
-    "examples/plot-caps-osc.jl",
-    "test/SedimentStackSpec.jl",
-    "src/BoundaryTrait.jl",
-    "test/Unitful.jl",
-    "examples/transport/active-layer-erosion.jl",
-    "examples/alcaps/plot-alternative.jl",
-    "examples/ca/long-term.jl",
-    "examples/production-only/plot-cap-slope.jl",
-    "examples/transport/active-layer-plot-result.jl",
-    "examples/transport/active-layer.jl",
-    "examples/alcaps/defaults.jl",
-    "examples/alcaps/plot-defaults.jl",
-    "src/CaProd.jl",
-    ".entangled/tasks.json",
-    "examples/alcaps/alternative.jl",
-    "examples/burgess2013-fig2.gnuplot",
-    "src/Erosion.jl",
-    "examples/transport/active-layer-plot-production.jl",
-    "examples/ca/life.jl",
-    "examples/production-only/ca-uniform.jl",
-    "src/Burgess2013/CA.jl",
-    "test/CASpec.jl",
-    "src/Transport/ActiveLayer.jl",
-    "examples/BS92/BS92.jl",
     "examples/ca/burgess-2013.jl",
-    "src/Burgess2013/Production.jl",
-    "src/Model/ALCAPS.jl",
-    "examples/plot-tanh.gnuplot",
-    "examples/BS92/fig8.jl",
-    "src/SedimentStack.jl",
-    "ext/VisualizationExt.jl",
-    "examples/production-only/ca-slope.jl",
+    "test/CASpec.jl",
+    "src/Visualization.jl",
+    "examples/transport/active-layer-plot-result.jl",
     "src/Stencil.jl",
-    "examples/production-only/caps-osc.jl",
     "src/Burgess2013.jl",
     "src/Vectors.jl",
-    "src/Config.jl",
-    "src/Burgess2013/Config.jl",
-    "src/Burgess2013/Transport.jl",
-    "examples/ca/eca.jl",
-    "src/Utility.jl",
+    "examples/burgess2013-fig2.gnuplot",
+    "src/SedimentStack.jl",
+    "examples/alcaps/plot-defaults.jl",
+    "examples/BS92/BS92.jl",
+    "examples/alcaps/alternative.jl",
+    "examples/production-only/ca-uniform.jl",
+    "examples/plot-tanh.gnuplot",
     "examples/BS92/fig8-sealevel.jl",
-    "test/ConfigSpec.jl"
+    "src/Transport/ActiveLayer.jl",
+    "examples/transport/active-layer.jl",
+    "examples/alcaps/plot-alternative.jl",
+    "src/Burgess2013/Production.jl",
+    "test/ConfigSpec.jl",
+    "src/Burgess2013/Config.jl",
+    "examples/ca/life.jl",
+    "src/Model/ALCAPS.jl",
+    "examples/transport/active-layer-erosion.jl",
+    "src/Burgess2013/CA.jl",
+    "src/BoundaryTrait.jl",
+    "src/Config.jl",
+    "src/Burgess2013/Transport.jl",
+    "src/Utility.jl",
+    "test/Unitful.jl",
+    "examples/production-only/ca-slope.jl",
+    "examples/BS92/fig8.jl",
+    "examples/transport/active-layer-plot-production.jl",
+    "test/SedimentStackSpec.jl",
+    "examples/plot-caps-osc.jl",
+    "ext/VisualizationExt.jl",
+    "examples/alcaps/defaults.jl",
+    "examples/ca/long-term.jl",
+    "examples/production-only/plot-cap-slope.jl",
+    "examples/ca/eca.jl",
+    "examples/production-only/caps-osc.jl",
+    "src/CaProd.jl",
+    ".entangled/tasks.json"
   ]
 }
