+++ src/CarboKitten.jl
@@ -5,18 +5,11 @@ include("./Vectors.jl")
 include("./Config.jl")
 include("./Stencil.jl")
 include("./SedimentStack.jl")
-include("./Burgess2013.jl")
 include("./Utility.jl")
-#include("./BS92.jl")
-include("./InputConfig.jl")
+include("./Burgess2013.jl")
+
 include("./Denudation.jl")
-
 include("./CaProd.jl")
-include("./CaProdErosion.jl")
-
-#include("./Visualization.jl")
-
-
 include("./Visualization.jl")
 
 module Transport
@@ -25,6 +18,7 @@ end
 
 module Model
 include("./Model/ALCAPS.jl")
+include("./Model/WithDenudation.jl")
 end
 
 end # module CarboKitten
